chore: move components folder outside of app folder
