+++ nextjs-app/components/pages/main/StayUpdated.tsx
